+++ engine/shared/library/sharedPathfinding/src/shared/PathSearch.cpp
@@ -182,7 +182,7 @@ PathSearchNode * PathSearchNode::createSearchNode( PathNode const * node )
 
 	PathSearchNode * oldNode = nullptr;
 
-	int mark = node->getMark(3);
+	int64_t mark = node->getMark(3);
 
 	if(mark != -1)
 	{
@@ -193,7 +193,7 @@ PathSearchNode * PathSearchNode::createSearchNode( PathNode const * node )
 
 	PathSearchNode * searchNode = new PathSearchNode(m_search,m_graph,node);
 
-	node->setMark( 3, (int)((void*)searchNode) );
+	node->setMark( 3, (int64_t)((void*)searchNode) );
 
 	m_search->m_visitedNodes->push_back(node);
